problema5.c

+++ C/Laboratorio4/problema3.c
@@ -13,33 +13,33 @@ Entradas, Salidas, Resumen:
 //Declaración e inicialización de variables globales
 
 //Llamamos a las funciones correspondientes.
-double norma(double V[]);
-void suma_vectores(double A[], double B[], int n);
-double producto_escalar(double A[], double B[], int n);
-void producto_vectorial(double A[], double B[], int n);
+float norma(float V[]);
+void suma_vectores(float A[], float B[], int n);
+float producto_escalar(float A[], float B[], int n);
+void producto_vectorial(float A[], float B[], int n);
 
 
 void main(){
     //Inicializamos los arreglos en los que vamos a guardar las coordenadas que ingrese el usuario.
     int n=3;
-    double A[3];
-    double B[3];
+    float A[3];
+    float B[3];
     
 
     //Solicita los valores a pares de las coordenadas de los dos vectores.
     for (int k; k < n; k++){
         printf("Ingrese las cordenadas %d  de los vectores A, B \n", k);
-        scanf("%lf, %lf", &A[k], &B[k]);
+        scanf("%f, %f", &A[k], &B[k]);
     }
     //Imprime la norma de cada vector
     puts("Norma de A, Norma de B");
-    printf("%lf, %lf \n", norma(A), norma(B));
+    printf("%f, %f \n", norma(A), norma(B));
     //Imprime la suma de los vectores
     puts("Suma Vectorial de A + B:");
     suma_vectores(A,B,n);
     //Imprime el producto escalar
     puts("Producto escalar:");
-    printf("%lf \n", producto_escalar(A,B,3));
+    printf("%f \n", producto_escalar(A,B,3));
     //Imprime el producto cruz
     puts("Producto vectorial:");
     producto_vectorial(A,B,3);
@@ -48,8 +48,8 @@ void main(){
 
 
 //Magnitud del vector
-double norma(double V[]){
-    double suma_cuadrados;
+float norma(float V[]){
+    float suma_cuadrados;
     //Se suma los cuadrados de las coordenadas
     suma_cuadrados = pow(V[0], 2) + pow(V[1], 2) + pow(V[2], 2);
     //Regresa la raiz cuadrada de lo anterior.
@@ -57,8 +57,8 @@ double norma(double V[]){
 }
 
 //Suma vectorial
-void suma_vectores(double A[], double B[], int n){
-    double S[n];
+void suma_vectores(float A[], float B[], int n){
+    float S[n];
 
     //Con el for sumamos coordenada con coordenada.
     for(int i=0; i < n; i++){
@@ -66,15 +66,15 @@ void suma_vectores(double A[], double B[], int n){
     }
     //Imprimimos las coordenadas del nuevo vector
     for (int i=0; i < n; i++){
-        printf("%lf ", S[i]);
+        printf("%f ", S[i]);
     }
     printf("\n");
 }
 
 
 //Producto escalar
-double producto_escalar(double A[], double B[], int n){
-    double suma_producto;
+float producto_escalar(float A[], float B[], int n){
+    float suma_producto;
     //Con el for sumamos la multiplicación de coordenada con coordenada
     for(int i =0; i < n; i++){
         suma_producto += A[i]*B[i];
@@ -83,8 +83,8 @@ double producto_escalar(double A[], double B[], int n){
 }
 
 //Producto vectorial
-void producto_vectorial(double A[], double B[], int n){
-    double C[n];
+void producto_vectorial(float A[], float B[], int n){
+    float C[n];
     int p, q;
     //Dado que la posición de las coordenadas en la fórmula se repite cíclicamente se puede aprovechar esta propiedad y usar m\'odulos.
     for(int m = 0; m < n; m++){
@@ -94,7 +94,7 @@ void producto_vectorial(double A[], double B[], int n){
     }
     //Imprime las coordenadas del nuevo vector.
     for (int i = 0; i < n; i++){
-        printf("%lf ", C[i]);
+        printf("%f ", C[i]);
     }
     printf("\n");
 }
